Add an authorization feature

+++ frontend/src/components/DayDetails.js
@@ -1,14 +1,22 @@
 import { useDaysContext } from "../hooks/useDaysContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 // date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow';
 
 const DayDetails = ({ day }) => {
     const { dispatch } = useDaysContext();
+    const { user } = useAuthContext();
 
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('/days/' + day._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         });
         const json = await response.json();
